+++ JayYi/ArchitectureStudy/app/src/main/java/com/jay/architecturestudy/ui/kin/KinFragment.kt
@@ -6,6 +6,7 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import com.jay.architecturestudy.R
 import com.jay.architecturestudy.data.model.Kin
 import com.jay.architecturestudy.data.model.ResponseNaverQuery
+import com.jay.architecturestudy.data.repository.NaverSearchRepositoryImpl
 import com.jay.architecturestudy.network.Api
 import com.jay.architecturestudy.ui.BaseFragment
 import kotlinx.android.synthetic.main.fragemnt_movie.*
@@ -16,6 +17,10 @@ import retrofit2.Response
 class KinFragment : BaseFragment(R.layout.fragemnt_kin) {
 
     private lateinit var kinAdapter: KinAdapter
+
+    private val naverSearchRepository by lazy {
+        NaverSearchRepositoryImpl()
+    }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
@@ -38,22 +43,14 @@ class KinFragment : BaseFragment(R.layout.fragemnt_kin) {
     }
 
     override fun search(keyword: String) {
-        Api.getKin(keyword)
-            .enqueue(object : Callback<ResponseNaverQuery<Kin>> {
-                override fun onFailure(call: Call<ResponseNaverQuery<Kin>>, t: Throwable) {
-                    Log.e("Kin", "error=${t.message}")
-                }
-
-                override fun onResponse(
-                    call: Call<ResponseNaverQuery<Kin>>,
-                    response: Response<ResponseNaverQuery<Kin>>
-                ) {
-                    if (response.isSuccessful) {
-                        val body = response.body() ?: return
-                        kinAdapter.setData(body.items)
-                    }
-                }
-
-            })
+        naverSearchRepository.getKin(
+            keyword = keyword,
+            success = { responseKin ->
+                kinAdapter.setData(responseKin.kins)
+            },
+            fail = {e ->
+                Log.e("Kin", "error=${e.message}")
+            }
+        )
     }
 }
